+++ web/usuaris/src/__mocks__/handlers.ts
@@ -1,5 +1,28 @@
-// import { rest } from "msw";
+import { HttpResponse, http } from 'msw'
+import { urls } from '../constants'
 
-export const handlers = [];
+export const handlers = [
+  http.get(urls.getUsers, () =>
+    HttpResponse.json(
+      [
+        {
+          id: 'TestId',
+          name: 'Test Name',
+          status: 'INACTIVE',
+          createdAt: '2023-11-15T15:36:02.234Z',
+          itineraryName: 'backend-node',
+        },
+      ],
+      { status: 200 }
+    )
+  ),
+]
 
-export const errorHandlers = [];
+export const errorHandlers = [
+  http.get(urls.getUsers, () =>
+    HttpResponse.json({ message: 'Invalid Credentials' }, { status: 401 })
+  ),
+  http.get(urls.getUsers, () =>
+    HttpResponse.json({ message: 'Database error' }, { status: 500 })
+  ),
+]